requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.16.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-admin==1.0.4
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
makefun==1.15.6
MarkupSafe==3.0.2
pendulum==3.0.0
pwdlib==0.2.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.9.0
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
time-machine==2.16.0
tortoise-orm==0.22.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
